+++ app/controllers/aip/v1/items_controller.rb
@@ -1,0 +1,16 @@
+class Aip::V1::ItemsController < ApplicationController
+
+  before_action :load_item, only: [:show]
+
+  def show
+    authorize @item
+    render json: { item: @item }, status: :ok
+  end
+
+  private
+
+  def load_item
+    @item = Item.find(params[:id])
+  end
+
+end
