--- app/controllers/api/v1/items_controller.rb
@@ -1,11 +1,0 @@
-class Api::V1::ItemsController < ApplicationController
-  include ItemLoad
-
-  before_action :load_item, only: [:show]
-
-  def show
-    authorize @item    
-    render json: {status: 'SUCCESS', data: @item}, status: :ok
-  end
-
-end
