+++ app/controllers/items_controller.rb
@@ -1,7 +1,5 @@
 class ItemsController < ApplicationController
 
-  include ItemLoad
-
   after_action :update_item_statistics, only: :show, unless: -> { request.bot? }
   before_action :load_item, only: [:show, :edit]
 
@@ -25,6 +23,14 @@ class ItemsController < ApplicationController
   end
 
   private
+
+  def load_item
+    @item = begin
+      Item.find(params[:id])
+    rescue ActiveRecord::RecordNotFound
+      Thesis.find(params[:id])
+    end
+  end
 
   def update_item_statistics
     Statistics.increment_view_count_for(item_id: params[:id], ip: request.ip)
