+++ config/routes.rb
@@ -116,11 +116,10 @@ Rails.application.routes.draw do
 
   match '/oai/(*all)', to: 'application#service_unavailable', via: [:get, :post]
 
-  # API v1
-  namespace :api, defaults: {format: :json} do
+  # AIP v1
+  namespace :aip, defaults: { format: :json } do
     namespace :v1 do
       resources :items, only: [:show]
     end
   end
-  
 end
