+++ test/fixtures/collections.yml
@@ -1,0 +1,4 @@
+desolate_collection:
+  title: Desolate collection
+  community: desolate_community
+  owner: admin
