+++ test/fixtures/communities.yml
@@ -1,0 +1,3 @@
+desolate_community:
+  title: Desolate community
+  owner: admin
